Retain causal formation components in f64
Preserve all nine final-state causes and prevent wire precision from feeding any scientific kernel.

+++ tests/formation_hillslope.rs
@@ -21,7 +21,7 @@ fn surface(radius_m: f64, target_cell_count: u32) -> SphericalSurfaceSnapshot {
 }
 
 struct Fields {
-    elevation_m: Vec<f32>,
+    elevation_m: Vec<f64>,
     water: SurfaceWaterField,
     erodibility: Vec<f32>,
     fracture: Vec<f32>,
@@ -48,7 +48,7 @@ impl Fields {
     }
 }
 
-fn uniform_fields(count: usize, elevation_m: f32) -> Fields {
+fn uniform_fields(count: usize, elevation_m: f64) -> Fields {
     Fields {
         elevation_m: vec![elevation_m; count],
         water: SurfaceWaterField::from_kinds(vec![SurfaceWaterKind::DryLand; count]),
@@ -135,7 +135,7 @@ fn isolated_edge_fields(
     let count = surface.cells().len();
     let (high, low, distance_m) = first_edge_cells(surface);
     let mut fields = uniform_fields(count, 0.0);
-    fields.elevation_m[high.raw() as usize] = (slope * distance_m) as f32;
+    fields.elevation_m[high.raw() as usize] = slope * distance_m;
     let mut water = vec![SurfaceWaterKind::Ocean; count];
     water[high.raw() as usize] = SurfaceWaterKind::DryLand;
     water[low.raw() as usize] = SurfaceWaterKind::DryLand;
@@ -233,12 +233,12 @@ fn finite_volume_cfl_rejects_an_unstable_step_instead_of_clipping_flux() {
         .elevation_m
         .iter()
         .copied()
-        .fold(f32::INFINITY, f32::min);
+        .fold(f64::INFINITY, f64::min);
     let input_maximum = fields
         .elevation_m
         .iter()
         .copied()
-        .fold(f32::NEG_INFINITY, f32::max);
+        .fold(f64::NEG_INFINITY, f64::max);
     assert!(stable
         .elevation_m()
         .iter()
@@ -270,8 +270,8 @@ fn normalized_edge_flux_is_resolution_invariant_within_the_monotone_step() {
             .iter()
             .find(|edge| edge.cells.contains(&high) && edge.cells.contains(&low))
             .unwrap();
-        let retained_slope = (f64::from(fields.elevation_m[high.raw() as usize])
-            - f64::from(fields.elevation_m[low.raw() as usize]))
+        let retained_slope = (fields.elevation_m[high.raw() as usize]
+            - fields.elevation_m[low.raw() as usize])
             / edge.center_distance.get();
         result.removed_volume_m3() / (0.001 * edge.length.get() * retained_slope)
     };
